requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.1.1
cryptography==43.0.1
python-dotenv==1.0.1
pytest==8.3.3
psycopg[binary]==3.2.3
